Mod: "chat, room api 분리"
chat, room api 분리

Resolves: N/A
Ref: N/A
Related to: N/A
